Scenario Entity integration
