+++ src/Managers/ScenarioManager.php
@@ -36,12 +36,13 @@ namespace NextDom\Managers;
 use NextDom\Helpers\FileSystemHelper;
 use NextDom\Helpers\Utils;
 use NextDom\Helpers\NextDomHelper;
+use NextDom\Model\Entity\Scenario;
 
 // TODO: \DB::buildField(ScenarioEntity::className) à factoriser
 class ScenarioManager
 {
     const DB_CLASS_NAME = 'scenario';
-    const CLASS_NAME = 'scenario';
+    const CLASS_NAME = Scenario::class;
     const INITIAL_TRANSLATION_FILE = '';
 
     /**
@@ -49,7 +50,7 @@ class ScenarioManager
      *
      * @param int $id Identifiant du scénario
      *
-     * @return \scenario Objet demandé
+     * @return Scenario Objet demandé
      *
      * @throws \Exception
      */
@@ -66,7 +67,7 @@ class ScenarioManager
      * @param string $scenarioName Chaine identifiant le scénario
      *
      * @param $commandNotFoundString
-     * @return \scenario Objet demandé
+     * @return Scenario Objet demandé
      *
      * @throws \ReflectionException
      */
@@ -85,7 +86,7 @@ class ScenarioManager
      * @param string $groupName Filtrer sur un groupe
      * @param string $type Filtrer sur un type
      *
-     * @return  \scenario[] Liste des objets scenario
+     * @return  Scenario[] Liste des objets scenario
      * @throws \Exception
      */
     public static function all($groupName = '', $type = null):array
