+++ src/Model/Entity/Scenario.php
@@ -17,11 +17,26 @@
 
 namespace NextDom\Model\Entity;
 
+use NextDom\Exceptions\CoreException;
+use NextDom\Helpers\NextDomHelper;
+use NextDom\Helpers\SystemHelper;
+use NextDom\Helpers\TimeLine;
+use NextDom\Helpers\Utils;
+use NextDom\Managers\CacheManager;
+use NextDom\Managers\CmdManager;
+use NextDom\Managers\ConfigManager;
+use NextDom\Managers\CronManager;
+use NextDom\Managers\DataStoreManager;
+use NextDom\Managers\EqLogicManager;
+use NextDom\Managers\EventManager;
+use NextDom\Managers\ScenarioElementManager;
+use NextDom\Managers\ScenarioManager;
+
 /**
  * Scenario
  *
- * @ORM\Table(name="scenario", uniqueConstraints={@ORM\UniqueConstraint(name="name", columns={"group", "object_id", "name"})}, indexes={@ORM\Index(name="group", columns={"group"}), @ORM\Index(name="fk_scenario_object1_idx", columns={"object_id"}), @ORM\Index(name="trigger", columns={"trigger"}), @ORM\Index(name="mode", columns={"mode"}), @ORM\Index(name="modeTriger", columns={"mode", "trigger"})})
- * @ORM\Entity
+ * ORM\Table(name="scenario", uniqueConstraints={@ORM\UniqueConstraint(name="name", columns={"group", "object_id", "name"})}, indexes={@ORM\Index(name="group", columns={"group"}), @ORM\Index(name="fk_scenario_object1_idx", columns={"object_id"}), @ORM\Index(name="trigger", columns={"trigger"}), @ORM\Index(name="mode", columns={"mode"}), @ORM\Index(name="modeTriger", columns={"mode", "trigger"})})
+ * ORM\Entity
  */
 class Scenario
 {
@@ -31,91 +46,91 @@ class Scenario
      *
      * @ORM\Column(name="name", type="string", length=127, nullable=true)
      */
-    private $name;
+    protected $name;
 
     /**
      * @var string
      *
      * @ORM\Column(name="group", type="string", length=127, nullable=true)
      */
-    private $group;
+    protected $group = '';
 
     /**
      * @var boolean
      *
      * @ORM\Column(name="isActive", type="boolean", nullable=true)
      */
-    private $isactive = '1';
+    protected $isActive = 1;
 
     /**
      * @var string
      *
      * @ORM\Column(name="mode", type="string", length=127, nullable=true)
      */
-    private $mode;
+    protected $mode;
 
     /**
      * @var string
      *
      * @ORM\Column(name="schedule", type="text", length=65535, nullable=true)
      */
-    private $schedule;
+    protected $schedule;
 
     /**
      * @var string
      *
      * @ORM\Column(name="scenarioElement", type="text", length=65535, nullable=true)
      */
-    private $scenarioelement;
+    protected $scenarioElement;
 
     /**
-     * @var string
+     * @var mixed
      *
      * @ORM\Column(name="trigger", type="string", length=255, nullable=true)
      */
-    private $trigger;
+    protected $trigger;
 
     /**
      * @var integer
      *
      * @ORM\Column(name="timeout", type="integer", nullable=true)
      */
-    private $timeout;
+    protected $timeout = 0;
 
     /**
      * @var boolean
      *
      * @ORM\Column(name="isVisible", type="boolean", nullable=true)
      */
-    private $isvisible = '1';
+    protected $isVisible = 1;
 
     /**
      * @var string
      *
      * @ORM\Column(name="display", type="text", length=65535, nullable=true)
      */
-    private $display;
+    protected $display;
 
     /**
      * @var string
      *
      * @ORM\Column(name="description", type="text", length=65535, nullable=true)
      */
-    private $description;
+    protected $description;
 
     /**
      * @var string
      *
      * @ORM\Column(name="configuration", type="text", length=65535, nullable=true)
      */
-    private $configuration;
+    protected $configuration;
 
     /**
      * @var string
      *
      * @ORM\Column(name="type", type="string", length=127, nullable=true)
      */
-    private $type = 'expert';
+    protected $type = 'expert';
 
     /**
      * @var integer
@@ -124,18 +139,27 @@ class Scenario
      * @ORM\Id
      * @ORM\GeneratedValue(strategy="IDENTITY")
      */
-    private $id;
+    protected $id;
 
     /**
      * @var \NextDom\Model\Entity\Object
      *
-     * @ORM\ManyToOne(targetEntity="NextDom\Model\Entity\Object")
-     * @ORM\JoinColumns({
-     *   @ORM\JoinColumn(name="object_id", referencedColumnName="id")
+     * ORM\ManyToOne(targetEntity="NextDom\Model\Entity\Object")
+     * ORM\JoinColumns({
+     *   ORM\JoinColumn(name="object_id", referencedColumnName="id")
      * })
      */
-    private $object;
+    protected $object_id;
 
+    protected static $_templateArray;
+    protected $_elements = array();
+    protected $_changeState = false;
+    protected $_realTrigger = '';
+    protected $_return = true;
+    protected $_tags = array();
+    protected $_do = true;
+    protected $_log;
+    
     public function getName()
     {
         return $this->name;
@@ -146,9 +170,9 @@ class Scenario
         return $this->group;
     }
 
-    public function getIsactive()
+    public function getIsActive()
     {
-        return $this->isactive;
+        return $this->isActive;
     }
 
     public function getMode()
@@ -158,32 +182,62 @@ class Scenario
 
     public function getSchedule()
     {
-        return $this->schedule;
+        return Utils::isJson($this->schedule, $this->schedule);
     }
 
-    public function getScenarioelement()
+    /**
+     *
+     * @return mixed
+     */
+    public function getScenarioElement()
     {
-        return $this->scenarioelement;
+        return Utils::isJson($this->scenarioElement, $this->scenarioElement);
     }
 
+    /**
+     *
+     * @return mixed
+     */
     public function getTrigger()
     {
-        return $this->trigger;
+        return Utils::isJson($this->trigger, array($this->trigger));
     }
 
-    public function getTimeout()
+    /**
+     *
+     * @param mixed $_default
+     * @return mixed
+     */
+    public function getTimeout($_default = null)
     {
+        if ($this->timeout == '' || !is_numeric($this->timeout)) {
+            return $_default;
+        }
         return $this->timeout;
     }
 
-    public function getIsvisible()
+    /**
+     *
+     * @param mixed $defaultValue
+     * @return mixed
+     */
+    public function getIsVisible($defaultValue = 0)
     {
-        return $this->isvisible;
+        if ($this->isVisible == '' || !is_numeric($this->isVisible)) {
+            return $defaultValue;
+        }
+        return $this->isVisible;
     }
 
-    public function getDisplay()
+    /**
+     *
+     * @param mixed $_key
+     * @param mixed $_default
+     * @return mixed
+     */
+    public function getDisplay($_key = '', $_default = '')
     {
-        return $this->display;
+        return Utils::getJsonAttr($this->display, $_key, $_default);
     }
 
     public function getDescription()
@@ -191,9 +245,29 @@ class Scenario
         return $this->description;
     }
 
-    public function getConfiguration()
+    /**
+     *
+     * @param string $key
+     * @param mixed $defaultValue
+     *
+     * @return mixed
+     */
+    public function getConfiguration($key = '', $defaultValue = '')
     {
-        return $this->configuration;
+        return Utils::getJsonAttr($this->configuration, $key, $defaultValue);
+    }
+
+    /**
+     *
+     * @param string $key
+     * @param mixed $value
+     *
+     * @return $this
+     */
+    public function setConfiguration($key, $value)
+    {
+        $this->configuration = Utils::setJsonAttr($this->configuration, $key, $value);
+        return $this;
     }
 
     public function getType()
@@ -206,26 +280,52 @@ class Scenario
         return $this->id;
     }
 
-    public function getObject(): \NextDom\Model\Entity\Object
+    /**
+     *
+     * @param mixed $_default
+     * @return mixed
+     */
+    public function getObject_id($_default = null)
     {
-        return $this->object;
+        if ($this->object_id == '' || !is_numeric($this->object_id)) {
+            return $_default;
+        }
+        return $this->object_id;
     }
 
     public function setName($name)
     {
+        if ($name != $this->getName()) {
+            $this->_changeState = true;
+        }
         $this->name = $name;
         return $this;
     }
 
+    /**
+     *
+     * @param mixed $group
+     * @return $this
+     */
     public function setGroup($group)
     {
+        if ($group != $this->getGroup()) {
+            $this->_changeState = true;
+        }
         $this->group = $group;
         return $this;
     }
-
-    public function setIsactive($isactive)
+    /**
+     *
+     * @param int $isActive
+     * @return $this
+     */
+    public function setIsActive($isActive)
     {
-        $this->isactive = $isactive;
+        if ($isActive != $this->getIsActive()) {
+            $this->_changeState = true;
+        }
+        $this->isActive = $isActive;
         return $this;
     }
 
@@ -237,49 +337,77 @@ class Scenario
 
     public function setSchedule($schedule)
     {
+        if (is_array($schedule)) {
+            $schedule = json_encode($schedule, JSON_UNESCAPED_UNICODE);
+        }
         $this->schedule = $schedule;
         return $this;
     }
 
-    public function setScenarioelement($scenarioelement)
+    /**
+     *
+     * @param mixed $scenarioElement
+     * @return $this
+     */
+    public function setScenarioElement($scenarioElement)
     {
-        $this->scenarioelement = $scenarioelement;
+        if (is_array($scenarioElement)) {
+            $scenarioElement = json_encode($scenarioElement, JSON_UNESCAPED_UNICODE);
+        }
+        $this->scenarioElement = $scenarioElement;
         return $this;
     }
 
+    /**
+     *
+     * @param mixed $trigger
+     * @return $this
+     * @throws \Exception
+     */
     public function setTrigger($trigger)
     {
-        $this->trigger = $trigger;
+        if (is_array($trigger)) {
+            $trigger = json_encode($trigger, JSON_UNESCAPED_UNICODE);
+        }
+        $this->trigger = CmdManager::humanReadableToCmd($trigger);
         return $this;
     }
 
+    /**
+     *
+     * @param string $timeout
+     * @return $this
+     */
     public function setTimeout($timeout)
     {
+        if ($timeout == '' || is_nan(intval($timeout)) || $timeout < 1) {
+            $timeout = '';
+        }
         $this->timeout = $timeout;
         return $this;
     }
 
-    public function setIsvisible($isvisible)
+    public function setIsVisible($isVisible)
     {
-        $this->isvisible = $isvisible;
+        $this->isVisible = $isVisible;
         return $this;
     }
 
-    public function setDisplay($display)
+    /**
+     *
+     * @param mixed $_key
+     * @param mixed $_value
+     * @return $this
+     */
+    public function setDisplay($_key, $_value)
     {
-        $this->display = $display;
+        $this->display = Utils::setJsonAttr($this->display, $_key, $_value);
         return $this;
     }
 
     public function setDescription($description)
     {
         $this->description = $description;
-        return $this;
-    }
-
-    public function setConfiguration($configuration)
-    {
-        $this->configuration = $configuration;
         return $this;
     }
 
@@ -295,10 +423,1130 @@ class Scenario
         return $this;
     }
 
-    public function setObject(\NextDom\Model\Entity\Object $object)
+    /**
+     *
+     * @param mixed $object_id
+     * @return $this
+     */
+    public function setObject_id($object_id = null)
     {
-        $this->object = $object;
+        if ($object_id != $this->getObject_id()) {
+            $this->_changeState = true;
+        }
+        $this->object_id = (!is_numeric($object_id)) ? null : $object_id;
         return $this;
+    }
+
+    /**
+     *
+     * @param mixed $_event
+     * @return boolean
+     */
+    public function testTrigger($_event)
+    {
+        foreach ($this->getTrigger() as $trigger) {
+            $trigger = str_replace(array('#variable(', ')#'), array('variable(', ')'), $trigger);
+            if ($trigger == $_event) {
+                return true;
+            } elseif (strpos($trigger, $_event) !== false && NextDomHelper::evaluateExpression($trigger)) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    /**
+     * Lance un scénario
+     *
+     * @param string $trigger
+     * @param string $message
+     * @param bool $forceSyncMode Force synchronous mode
+     * @return boolean
+     * @throws \Exception
+     */
+    public function launch($trigger = '', $message = '', bool $forceSyncMode = false)
+    {
+        // Test if scenarios are enabled and if this scenario is activated
+        if (ConfigManager::byKey('enableScenario') != 1 || $this->getIsActive() != 1) {
+            return false;
+        }
+        // Test execution mode
+        if ($this->getConfiguration('syncmode') == 1 || $forceSyncMode) {
+            $this->setLog(__('Lancement du scénario en mode synchrone'));
+            return $this->execute($trigger, $message);
+        } else {
+            if (count($this->getTags()) != '') {
+                $this->setCache('tags', $this->getTags());
+            }
+            $cmd = NEXTDOM_ROOT . '/core/php/jeeScenario.php ';
+            $cmd .= ' scenario_id=' . $this->getId();
+            $cmd .= ' trigger=' . escapeshellarg($trigger);
+            $cmd .= ' "message=' . escapeshellarg(sanitizeAccent($message)) . '"';
+            $cmd .= ' >> ' . \log::getPathToLog('scenario_execution') . ' 2>&1 &';
+            SystemHelper::php($cmd);
+        }
+        return true;
+    }
+
+    /**
+     * Execute the scenario
+     *
+     * @param string $trigger
+     * @param string $message
+     * @return mixed
+     * @throws \Exception
+     */
+    public function execute($trigger = '', $message = '')
+    {
+        $tags = $this->getCache('tags');
+        if ($tags != '') {
+            $this->setTags($tags);
+            $this->setCache('tags', '');
+        }
+        if ($this->getIsActive() != 1) {
+            $this->setLog(__('Impossible d\'exécuter le scénario : ') . $this->getHumanName() . __(' sur : ') . $message . __(' car il est désactivé'));
+            $this->persistLog();
+            return null;
+        }
+        if ($this->getConfiguration('timeDependency', 0) == 1 && !NextDomHelper::isDateOk()) {
+            $this->setLog(__('Lancement du scénario : ') . $this->getHumanName() . __(' annulé car il utilise une condition de type temporelle et que la date système n\'est pas OK'));
+            $this->persistLog();
+            return null;
+        }
+
+        $cmd = CmdManager::byId(str_replace('#', '', $trigger));
+        if (is_object($cmd)) {
+            \log::add('event', 'info', __('Exécution du scénario ') . $this->getHumanName() . __(' déclenché par : ') . $cmd->getHumanName());
+            if ($this->getConfiguration('timeline::enable')) {
+                TimeLine::addTimelineEvent(array('type' => 'scenario', 'id' => $this->getId(), 'name' => $this->getHumanName(true), 'datetime' => date('Y-m-d H:i:s'), 'trigger' => $cmd->getHumanName(true)));
+            }
+        } else {
+            \log::add('event', 'info', __('Exécution du scénario ') . $this->getHumanName() . __(' déclenché par : ') . $trigger);
+            if ($this->getConfiguration('timeline::enable')) {
+                TimeLine::addTimelineEvent(array('type' => 'scenario', 'id' => $this->getId(), 'name' => $this->getHumanName(true), 'datetime' => date('Y-m-d H:i:s'), 'trigger' => $trigger == 'schedule' ? 'programmation' : $trigger));
+            }
+        }
+        if (count($this->getTags()) == 0) {
+            $this->setLog('Start : ' . trim($message, "'") . '.');
+        } else {
+            $this->setLog('Start : ' . trim($message, "'") . '. Tags : ' . json_encode($this->getTags()));
+        }
+        $this->setLastLaunch(date('Y-m-d H:i:s'));
+        $this->setState('in progress');
+        $this->setPID(getmypid());
+        $this->setRealTrigger($trigger);
+        foreach ($this->getElement() as $element) {
+            if (!$this->getDo()) {
+                break;
+            }
+            $element->execute($this);
+        }
+        $this->setState('stop');
+        $this->setPID();
+        $this->setLog(__('Fin correcte du scénario'));
+        $this->persistLog();
+        return $this->getReturn();
+    }
+
+    /**
+     *
+     * @param mixed $_name
+     * @return Scenario
+     * @throws \Exception
+     */
+    public function copy($_name)
+    {
+        $scenarioCopy = clone $this;
+        $scenarioCopy->setName($_name);
+        $scenarioCopy->setId('');
+        $scenario_element_list = array();
+        foreach ($this->getElement() as $element) {
+            $scenario_element_list[] = $element->copy();
+        }
+        $scenarioCopy->setScenarioElement($scenario_element_list);
+        $scenarioCopy->setLog('');
+        $scenarioCopy->save();
+        if (file_exists('/var/log/nextdom/scenarioLog/scenario' . $scenarioCopy->getId() . '.log')) {
+            unlink('/var/log/nextdom/scenarioLog/scenario' . $scenarioCopy->getId() . '.log');
+        }
+        return $scenarioCopy;
+    }
+
+    /**
+     *
+     * @param mixed $_version
+     * @return string
+     * @throws \Exception
+     */
+    public function toHtml($_version)
+    {
+        if (!$this->hasRight('r')) {
+            return '';
+        }
+        $mc = CacheManager::byKey('scenarioHtml' . $_version . $this->getId());
+        if ($mc->getValue() != '') {
+            return $mc->getValue();
+        }
+
+        $version = NextDomHelper::versionAlias($_version);
+        $replace = array(
+            '#id#' => $this->getId(),
+            '#state#' => $this->getState(),
+            '#isActive#' => $this->getIsActive(),
+            '#name#' => ($this->getDisplay('name') != '') ? $this->getDisplay('name') : $this->getHumanName(),
+            '#shortname#' => ($this->getDisplay('name') != '') ? $this->getDisplay('name') : $this->getName(),
+            '#treename#' => $this->getHumanName(false, false, false, false, true),
+            '#icon#' => $this->getIcon(),
+            '#lastLaunch#' => $this->getLastLaunch(),
+            '#scenarioLink#' => $this->getLinkToConfiguration(),
+            '#version#' => $_version,
+        );
+        if (!isset(self::$_templateArray)) {
+            self::$_templateArray = array();
+        }
+        if (!isset(self::$_templateArray[$version])) {
+            self::$_templateArray[$version] = getTemplate('core', $version, 'scenario');
+        }
+        $html = Utils::templateReplace($replace, self::$_templateArray[$version]);
+        CacheManager::set('scenarioHtml' . $version . $this->getId(), $html);
+        return $html;
+    }
+
+    /**
+     *
+     */
+    public function emptyCacheWidget()
+    {
+        $mc = CacheManager::byKey('scenarioHtmldashboard' . $this->getId());
+        $mc->remove();
+        $mc = CacheManager::byKey('scenarioHtmlmobile' . $this->getId());
+        $mc->remove();
+        $mc = CacheManager::byKey('scenarioHtmlmview' . $this->getId());
+        $mc->remove();
+        $mc = CacheManager::byKey('scenarioHtmldview' . $this->getId());
+        $mc->remove();
+    }
+
+    /**
+     *
+     * @param mixed $_only_class
+     * @return string
+     * @throws \Exception
+     */
+    public function getIcon($_only_class = false)
+    {
+        if ($_only_class) {
+            if ($this->getIsActive() == 1) {
+                switch ($this->getState()) {
+                    case 'in progress':
+                        return 'fas fa-spinner fa-spin';
+                    case 'error':
+                        return 'fas fa-exclamation-triangle';
+                    default:
+                        if (strpos($this->getDisplay('icon'), '<i') === 0) {
+                            return str_replace(array('<i', 'class=', '"', '/>'), '', $this->getDisplay('icon'));
+                        }
+                        return 'fas fa-check';
+                }
+            } else {
+                return 'fas fa-times';
+            }
+        } else {
+            if ($this->getIsActive() == 1) {
+                switch ($this->getState()) {
+                    case 'in progress':
+                        return '<i class="fas fa-spinner fa-spin"></i>';
+                    case 'error':
+                        return '<i class="fas fa-exclamation-triangle"></i>';
+                    default:
+                        if (strpos($this->getDisplay('icon'), '<i') === 0) {
+                            return $this->getDisplay('icon');
+                        }
+                        return '<i class="fas fa-check"></i>';
+                }
+            } else {
+                return '<i class="fas fa-times"></i>';
+            }
+        }
+    }
+
+    /**
+     *
+     * @return string
+     */
+    public function getLinkToConfiguration()
+    {
+        return 'index.php?v=d&p=scenario&id=' . $this->getId();
+    }
+
+    /**
+     *
+     * @throws CoreException
+     */
+    public function preSave()
+    {
+        if ($this->getTimeout() == '' || !is_numeric($this->getTimeout())) {
+            $this->setTimeout(0);
+        }
+        if ($this->getName() == '') {
+            throw new CoreException(__('Le nom du scénario ne peut pas être vide.'));
+        }
+        if (($this->getMode() == 'schedule' || $this->getMode() == 'all') && $this->getSchedule() == '') {
+            throw new CoreException(__('Le scénario est de type programmé mais la programmation est vide'));
+        }
+        if ($this->getConfiguration('has_return', 0) == 1) {
+            $this->setConfiguration('syncmode', 1);
+        }
+        if ($this->getConfiguration('logmode') == '') {
+            $this->setConfiguration('logmode', 'default');
+        }
+    }
+
+    /**
+     *
+     */
+    public function postInsert()
+    {
+        $this->setState('stop');
+        $this->setPID();
+    }
+
+    /**
+     *
+     */
+    public function save()
+    {
+        if ($this->getLastLaunch() == '' && ($this->getMode() == 'schedule' || $this->getMode() == 'all')) {
+            $calculateScheduleDate = $this->calculateScheduleDate();
+            $this->setLastLaunch($calculateScheduleDate['prevDate']);
+        }
+        \DB::save($this);
+        $this->emptyCacheWidget();
+        if ($this->_changeState) {
+            $this->_changeState = false;
+            EventManager::add('scenario::update', array('scenario_id' => $this->getId(), 'isActive' => $this->getIsActive(), 'state' => $this->getState(), 'lastLaunch' => $this->getLastLaunch()));
+        }
+    }
+
+    /**
+     *
+     */
+    public function refresh()
+    {
+        \DB::refresh($this);
+    }
+
+    /**
+     *
+     * @return mixed
+     * @throws \Exception
+     */
+    public function remove()
+    {
+        \viewData::removeByTypeLinkId('scenario', $this->getId());
+        \dataStore::removeByTypeLinkId('scenario', $this->getId());
+        foreach ($this->getElement() as $element) {
+            $element->remove();
+        }
+        $this->emptyCacheWidget();
+        if (file_exists('/var/log/nextdom/scenarioLog/scenario' . $this->getId() . '.log')) {
+            unlink('/var/log/nextdom/scenarioLog/scenario' . $this->getId() . '.log');
+        }
+        CacheManager::delete('scenarioCacheAttr' . $this->getId());
+        return \DB::remove($this);
+    }
+
+    /**
+     *
+     * @param mixed $_key
+     * @param mixed $_private
+     * @return boolean
+     * @throws \Exception
+     */
+    public function removeData($_key, $_private = false)
+    {
+        if ($_private) {
+            $dataStore = DataStoreManager::byTypeLinkIdKey('scenario', $this->getId(), $_key);
+        } else {
+            $dataStore = DataStoreManager::byTypeLinkIdKey('scenario', -1, $_key);
+        }
+        if (is_object($dataStore)) {
+            return $dataStore->remove();
+        }
+        return true;
+    }
+
+    /**
+     *
+     * @param mixed $_key
+     * @param mixed $_value
+     * @param bool $_private
+     * @return boolean
+     */
+    public function setData($_key, $_value, $_private = false)
+    {
+        $dataStore = new \dataStore();
+        $dataStore->setType('scenario');
+        $dataStore->setKey($_key);
+        $dataStore->setValue($_value);
+        if ($_private) {
+            $dataStore->setLink_id($this->getId());
+        } else {
+            $dataStore->setLink_id(-1);
+        }
+        $dataStore->save();
+        return true;
+    }
+
+    public function getData($_key, $_private = false, $_default = '')
+    {
+        if ($_private) {
+            $dataStore = DataStoreManager::byTypeLinkIdKey('scenario', $this->getId(), $_key);
+        } else {
+            $dataStore = DataStoreManager::byTypeLinkIdKey('scenario', -1, $_key);
+        }
+        if (is_object($dataStore)) {
+            return $dataStore->getValue($_default);
+        }
+        return $_default;
+    }
+
+    /**
+     *
+     * @return mixed
+     */
+    public function calculateScheduleDate()
+    {
+        $calculatedDate = array('prevDate' => '', 'nextDate' => '');
+        if (is_array($this->getSchedule())) {
+            $calculatedDate_tmp = array('prevDate' => '', 'nextDate' => '');
+            foreach ($this->getSchedule() as $schedule) {
+                try {
+                    $c = new \Cron\CronExpression($schedule, new \Cron\FieldFactory);
+                    $calculatedDate_tmp['prevDate'] = $c->getPreviousRunDate()->format('Y-m-d H:i:s');
+                    $calculatedDate_tmp['nextDate'] = $c->getNextRunDate()->format('Y-m-d H:i:s');
+                } catch (\Exception $exc) {
+
+                }
+                if ($calculatedDate['prevDate'] == '' || strtotime($calculatedDate['prevDate']) < strtotime($calculatedDate_tmp['prevDate'])) {
+                    $calculatedDate['prevDate'] = $calculatedDate_tmp['prevDate'];
+                }
+                if ($calculatedDate['nextDate'] == '' || strtotime($calculatedDate['nextDate']) > strtotime($calculatedDate_tmp['nextDate'])) {
+                    $calculatedDate['nextDate'] = $calculatedDate_tmp['nextDate'];
+                }
+            }
+        } else {
+            try {
+                $c = new \Cron\CronExpression($this->getSchedule(), new \Cron\FieldFactory);
+                $calculatedDate['prevDate'] = $c->getPreviousRunDate()->format('Y-m-d H:i:s');
+                $calculatedDate['nextDate'] = $c->getNextRunDate()->format('Y-m-d H:i:s');
+            } catch (\Exception $exc) {
+
+            }
+        }
+        return $calculatedDate;
+    }
+
+    /**
+     *
+     * @return boolean
+     * @throws \Exception
+     */
+    public function isDue()
+    {
+        $last = strtotime($this->getLastLaunch());
+        $now = time();
+        $now = ($now - $now % 60);
+        $last = ($last - $last % 60);
+        if ($now == $last) {
+            return false;
+        }
+        if (is_array($this->getSchedule())) {
+            foreach ($this->getSchedule() as $schedule) {
+                try {
+                    $c = new \Cron\CronExpression($schedule, new \Cron\FieldFactory);
+                    try {
+                        if ($c->isDue()) {
+                            return true;
+                        }
+                    } catch (\Exception $e) {
+
+                    }
+                    try {
+                        $prev = $c->getPreviousRunDate()->getTimestamp();
+                    } catch (\Exception $e) {
+                        continue;
+                    }
+                    $lastCheck = strtotime($this->getLastLaunch());
+                    $diff = abs((strtotime('now') - $prev) / 60);
+                    if ($lastCheck <= $prev && $diff <= ConfigManager::byKey('maxCatchAllow') || ConfigManager::byKey('maxCatchAllow') == -1) {
+                        return true;
+                    }
+                } catch (\Exception $e) {
+
+                }
+            }
+        } else {
+            try {
+                $c = new \Cron\CronExpression($this->getSchedule(), new \Cron\FieldFactory);
+                try {
+                    if ($c->isDue()) {
+                        return true;
+                    }
+                } catch (\Exception $e) {
+
+                }
+                try {
+                    $prev = $c->getPreviousRunDate()->getTimestamp();
+                } catch (\Exception $e) {
+                    return false;
+                }
+                $lastCheck = strtotime($this->getLastLaunch());
+                $diff = abs((strtotime('now') - $prev) / 60);
+                if ($lastCheck <= $prev && $diff <= ConfigManager::byKey('maxCatchAllow') || ConfigManager::byKey('maxCatchAllow') == -1) {
+                    return true;
+                }
+            } catch (\Exception $exc) {
+
+            } 
+        }
+        return false;
+    }
+
+    /**
+     *
+     * @return boolean
+     * @throws \Exception
+     */
+    public function running()
+    {
+        if (intval($this->getPID()) > 0 && posix_getsid(intval($this->getPID())) && (!file_exists('/proc/' . $this->getPID() . '/cmdline') || strpos(file_get_contents('/proc/' . $this->getPID() . '/cmdline'), 'scenario_id=' . $this->getId()) !== false)) {
+            return true;
+        }
+        if (count(SystemHelper::ps('scenario_id=' . $this->getId() . ' ', array(getmypid()))) > 0) {
+            return true;
+        }
+        return false;
+    }
+
+    /**
+     *
+     * @return boolean
+     * @throws \Exception
+     */
+    public function stop()
+    {
+        $crons = CronManager::searchClassAndFunction('scenario', 'doIn', '"scenario_id":' . $this->getId());
+        if (is_array($crons)) {
+            foreach ($crons as $cron) {
+                if ($cron->getState() == 'run') {
+                    try {
+                        $cron->halt();
+                        $cron->remove();
+                    } catch (\Exception $e) {
+                        \log::add('scenario', 'info', __('Can not stop subtask : ') . print_r($cron->getOption(), true));
+                    }
+                }
+            }
+        }
+        if ($this->running()) {
+            if ($this->getPID() > 0) {
+                SystemHelper::kill($this->getPID());
+                $retry = 0;
+                while ($this->running() && $retry < 10) {
+                    sleep(1);
+                    SystemHelper::kill($this->getPID());
+                    $retry++;
+                }
+            }
+
+            if ($this->running()) {
+                SystemHelper::kill("scenario_id=" . $this->getId() . ' ');
+                sleep(1);
+                if ($this->running()) {
+                    SystemHelper::kill("scenario_id=" . $this->getId() . ' ');
+                    sleep(1);
+                }
+            }
+            if ($this->running()) {
+                throw new CoreException(__('Impossible d\'arrêter le scénario : ') . $this->getHumanName() . __('. PID : ') . $this->getPID());
+            }
+        }
+        $this->setState('stop');
+        return true;
+    }
+
+    /**
+     *
+     * @return \scenarioElement[]
+     * @throws \Exception
+     */
+    public function getElement()
+    {
+        if (count($this->_elements) > 0) {
+            return $this->_elements;
+        }
+        $result = array();
+        $elements = $this->getScenarioElement();
+        $elementId = -1;
+        if (is_array($elements)) {
+            foreach ($this->getScenarioElement() as $elementId) {
+                $element = ScenarioElementManager::byId($elementId);
+                if (is_object($element)) {
+                    $result[] = $element;
+                }
+            }
+            $this->_elements = $result;
+            return $result;
+        }
+        if ($elements != '') {
+            $element = ScenarioElementManager::byId($elementId);
+            if (is_object($element)) {
+                $result[] = $element;
+                $this->_elements = $result;
+                return $result;
+            }
+        }
+        return array();
+    }
+
+    /**
+     *
+     * @param mixed $_mode
+     * @return mixed
+     * @throws \Exception
+     */
+    public function export($_mode = 'text')
+    {
+        if ($_mode == 'text') {
+            $return = '';
+            $return .= '- Nom du scénario : ' . $this->getName() . "\n";
+            if (is_numeric($this->getObject_id())) {
+                $return .= '- Objet parent : ' . $this->getObject()->getName() . "\n";
+            }
+            $return .= '- Mode du scénario : ' . $this->getMode() . "\n";
+            $schedules = $this->getSchedule();
+            if ($this->getMode() == 'schedule' || $this->getMode() == 'all') {
+                if (is_array($schedules)) {
+                    foreach ($schedules as $schedule) {
+                        $return .= '    - Programmation : ' . $schedule . "\n";
+                    }
+                } else {
+                    if ($schedules != '') {
+                        $return .= '    - Programmation : ' . $schedules . "\n";
+                    }
+                }
+            }
+            if ($this->getMode() == 'provoke' || $this->getMode() == 'all') {
+                foreach ($this->getTrigger() as $trigger) {
+                    $return .= '    - Evènement : ' . NextDomHelper::toHumanReadable($trigger) . "\n";
+                }
+            }
+            $return .= "\n";
+            $return .= $this->getDescription();
+            $return .= "\n\n";
+            foreach ($this->getElement() as $element) {
+                $exports = explode("\n", $element->export());
+                foreach ($exports as $export) {
+                    $return .= "    " . $export . "\n";
+                }
+            }
+        }
+        $return = [];
+        if ($_mode == 'array') {
+            $return = Utils::o2a($this);
+            $return['trigger'] = NextDomHelper::toHumanReadable($return['trigger']);
+            $return['elements'] = array();
+            foreach ($this->getElement() as $element) {
+                $return['elements'][] = $element->getAjaxElement('array');
+            }
+            if (isset($return['id'])) {
+                unset($return['id']);
+            }
+            if (isset($return['lastLaunch'])) {
+                unset($return['lastLaunch']);
+            }
+            if (isset($return['log'])) {
+                unset($return['log']);
+            }
+            if (isset($return['hlogs'])) {
+                unset($return['hlogs']);
+            }
+            if (isset($return['object_id'])) {
+                unset($return['object_id']);
+            }
+            if (isset($return['pid'])) {
+                unset($return['pid']);
+            }
+            if (isset($return['scenarioElement'])) {
+                unset($return['scenarioElement']);
+            }
+            if (isset($return['_templateArray'])) {
+                unset($return['_templateArray']);
+            }
+            if (isset($return['_templateArray'])) {
+                unset($return['_templateArray']);
+            }
+            if (isset($return['_changeState'])) {
+                unset($return['_changeState']);
+            }
+            if (isset($return['_realTrigger'])) {
+                unset($return['_realTrigger']);
+            }
+            if (isset($return['_templateArray'])) {
+                unset($return['_templateArray']);
+            }
+            if (isset($return['_elements'])) {
+                unset($return['_elements']);
+            }
+        }
+        return $return;
+    }
+
+    /**
+     *
+     * @return \object
+     */
+    public function getObject()
+    {
+        return \object::byId($this->object_id);
+    }
+
+    /**
+     *
+     * @param mixed $_complete
+     * @param mixed $_noGroup
+     * @param mixed $_tag
+     * @param mixed $_prettify
+     * @param mixed $_withoutScenarioName
+     * @return string
+     */
+    public function getHumanName($_complete = false, $_noGroup = false, $_tag = false, $_prettify = false, $_withoutScenarioName = false)
+    {
+        $name = '';
+        if (is_numeric($this->getObject_id()) && is_object($this->getObject())) {
+            $object = $this->getObject();
+            if ($_tag) {
+                if ($object->getDisplay('tagColor') != '') {
+                    $name .= '<span class="label" style="text-shadow : none;background-color:' . $object->getDisplay('tagColor') . ' !important;color:' . $object->getDisplay('tagTextColor', 'white') . ' !important">' . $object->getName() . '</span>';
+                } else {
+                    $name .= '<span class="label label-primary" style="text-shadow : none;">' . $object->getName() . '</span>';
+                }
+            } else {
+                $name .= '[' . $object->getName() . ']';
+            }
+        } else {
+            if ($_complete) {
+                if ($_tag) {
+                    $name .= '<span class="label label-default" style="text-shadow : none;">' . __('Aucun') . '</span>';
+                } else {
+                    $name .= '[' . __('Aucun') . ']';
+                }
+            }
+        }
+        if (!$_noGroup) {
+            if ($this->getGroup() != '') {
+                $name .= '[' . $this->getGroup() . ']';
+            } else {
+                if ($_complete) {
+                    $name .= '[' . __('Aucun') . ']';
+                }
+            }
+        }
+        if ($_prettify) {
+            $name .= '</p><p>';
+        }
+        if (!$_withoutScenarioName) {
+            if ($_tag) {
+                $name .= ' ' . $this->getName();
+            } else {
+                $name .= '[' . $this->getName() . ']';
+            }
+        }
+        if ($_prettify) {
+            $name .= '</p>';
+        }
+        return $name;
+    }
+
+    /**
+     *
+     * @param mixed $_right
+     * @param \user|null $_user
+     *
+     * @return boolean
+     */
+    public function hasRight($_right, $_user = null)
+    {
+        if ($_user !== null) {
+            if ($_user->getProfils() == 'admin' || $_user->getProfils() == 'user') {
+                return true;
+            }
+            if (strpos($_user->getRights('scenario' . $this->getId()), $_right) !== false) {
+                return true;
+            }
+            return false;
+        }
+        if (!isConnect()) {
+            return false;
+        }
+        if (isConnect('admin') || isConnect('user')) {
+            return true;
+        }
+        if (strpos($_SESSION['user']->getRights('scenario' . $this->getId()), $_right) !== false) {
+            return true;
+        }
+        return false;
+    }
+
+    /**
+     *
+     * @param mixed $_partial
+     * @return bool|null
+     */
+    public function persistLog($_partial = false)
+    {
+        if ($this->getConfiguration('logmode', 'default') == 'none') {
+            return null;
+        }
+        $path = '/var/log/nextdom/scenarioLog';
+        if (!file_exists($path)) {
+            mkdir($path);
+        }
+        $path .= '/scenario' . $this->getId() . '.log';
+        if ($_partial) {
+            file_put_contents($path, $this->getLog(), FILE_APPEND);
+        } else {
+            file_put_contents($path, "------------------------------------\n" . $this->getLog(), FILE_APPEND);
+        }
+        return true;
+    }
+
+    /**
+     *
+     * @return mixed
+     * @throws \Exception
+     */
+    public function toArray()
+    {
+        $return = Utils::o2a($this, true);
+        $cache = $this->getCache(array('state', 'lastLaunch'));
+        // TODO: Pourquoi ce test a-t-il dû être rajouté ?
+        if (isset($cache['state'])) {
+            $return['state'] = $cache['state'];
+        }
+        else {
+            $return['state'] = '';
+        }
+        if (isset($cache['lastLaunch'])) {
+            $return['lastLaunch'] = $cache['lastLaunch'];
+        }
+        else {
+            $return['lastLaunch'] = '';
+        }
+        return $return;
+    }
+
+    /**
+     *
+     * @param mixed $_data
+     * @param mixed $_level
+     * @param mixed $_drill
+     * @return string
+     * @throws \Exception
+     */
+    public function getLinkData(&$_data = array('node' => array(), 'link' => array()), $_level = 0, $_drill = null)
+    {
+        if ($_drill === null) {
+            $_drill = ConfigManager::byKey('graphlink::scenario::drill');
+        }
+        if (isset($_data['node']['scenario' . $this->getId()])) {
+            return null;
+        }
+        if ($this->getIsActive() == 0 && $_level > 0) {
+            return $_data;
+        }
+        $_level++;
+        if ($_level > $_drill) {
+            return $_data;
+        }
+
+        $_data['node']['scenario' . $this->getId()] = array(
+            'id' => 'scenario' . $this->getId(),
+            'name' => $this->getName(),
+            'fontweight' => ($_level == 1) ? 'bold' : 'normal',
+            'shape' => 'rect',
+            'width' => 40,
+            'height' => 40,
+            'color' => 'green',
+            'image' => '/public/img/NextDom_Scenario.png',
+            'title' => $this->getHumanName(),
+            'url' => 'index.php?v=d&p=scenario&id=' . $this->getId(),
+        );
+        $use = $this->getUse();
+        $usedBy = $this->getUsedBy();
+        Utils::addGraphLink($this, 'scenario', $this->getObject(), 'object', $_data, $_level + 1, $_drill, array('dashvalue' => '1,0', 'lengthfactor' => 0.6));
+        Utils::addGraphLink($this, 'scenario', $use['cmd'], 'cmd', $_data, $_level, $_drill);
+        Utils::addGraphLink($this, 'scenario', $use['scenario'], 'scenario', $_data, $_level, $_drill);
+        Utils::addGraphLink($this, 'scenario', $use['eqLogic'], 'eqLogic', $_data, $_level, $_drill);
+        Utils::addGraphLink($this, 'scenario', $use['dataStore'], 'dataStore', $_data, $_level, $_drill);
+        Utils::addGraphLink($this, 'scenario', $use['view'], 'view', $_data, $_level, $_drill);
+        Utils::addGraphLink($this, 'scenario', $use['plan'], 'plan', $_data, $_level, $_drill);
+        Utils::addGraphLink($this, 'scenario', $usedBy['cmd'], 'cmd', $_data, $_level, $_drill);
+        Utils::addGraphLink($this, 'scenario', $usedBy['scenario'], 'scenario', $_data, $_level, $_drill);
+        Utils::addGraphLink($this, 'scenario', $usedBy['eqLogic'], 'eqLogic', $_data, $_level, $_drill);
+        Utils::addGraphLink($this, 'scenario', $usedBy['interactDef'], 'interactDef', $_data, $_level, $_drill, array('dashvalue' => '2,6', 'lengthfactor' => 0.6));
+        Utils::addGraphLink($this, 'scenario', $usedBy['plan'], 'plan', $_data, $_level, $_drill, array('dashvalue' => '2,6', 'lengthfactor' => 0.6));
+        Utils::addGraphLink($this, 'scenario', $usedBy['view'], 'view', $_data, $_level, $_drill, array('dashvalue' => '2,6', 'lengthfactor' => 0.6));
+        return $_data;
+    }
+
+    /**
+     *
+     * @return mixed
+     * @throws \Exception
+     */
+    public function getUse()
+    {
+        $json = NextDomHelper::fromHumanReadable(json_encode($this->export('array')));
+        return NextDomHelper::getTypeUse($json);
+    }
+
+    /**
+     *
+     * @param bool $_array
+     * @return mixed
+     * @throws \Exception
+     */
+    public function getUsedBy($_array = false)
+    {
+        $return = array('cmd' => array(), 'eqLogic' => array(), 'scenario' => array(), 'plan' => array(), 'view' => array());
+        $return['cmd'] = CmdManager::searchConfiguration('#scenario' . $this->getId() . '#');
+        $return['eqLogic'] = EqLogicManager::searchConfiguration(array('#scenario' . $this->getId() . '#', '"scenario_id":"' . $this->getId()));
+        $return['interactDef'] = \interactDef::searchByUse(array('#scenario' . $this->getId() . '#', '"scenario_id":"' . $this->getId()));
+        $return['scenario'] = ScenarioManager::searchByUse(array(
+            array('action' => 'scenario', 'option' => $this->getId(), 'and' => true),
+            array('action' => '#scenario' . $this->getId() . '#'),
+        ));
+        $return['view'] = \view::searchByUse('scenario', $this->getId());
+        $return['plan'] = \planHeader::searchByUse('scenario', $this->getId());
+        if ($_array) {
+            foreach ($return as &$value) {
+                $value = Utils::o2a($value);
+            }
+        }
+        return $return;
+    }
+
+    public function clearLog()
+    {
+        $this->_log = '';
+    }
+
+    public function resetRepeatIfStatus() {
+        foreach ($this->getElement() as $element) {
+            $element->resetRepeatIfStatus();
+        }
+    }
+
+    /**
+     *
+     * @return mixed
+     * @throws \Exception
+     */
+    public function getState()
+    {
+        return $this->getCache('state');
+    }
+
+    /**
+     *
+     * @return mixed
+     * @throws \Exception
+     */
+    public function getLastLaunch()
+    {
+        return $this->getCache('lastLaunch');
+    }
+
+    /**
+     *
+     * @param mixed $state
+     * @throws \Exception
+     */
+    public function setState($state)
+    {
+        if ($this->getCache('state') != $state) {
+            $this->emptyCacheWidget();
+            EventManager::add('scenario::update', array('scenario_id' => $this->getId(), 'state' => $state, 'lastLaunch' => $this->getLastLaunch()));
+        }
+        $this->setCache('state', $state);
+    }
+
+    /**
+     *
+     * @param mixed $lastLaunch
+     * @throws \Exception
+     */
+    public function setLastLaunch($lastLaunch)
+    {
+        $this->setCache('lastLaunch', $lastLaunch);
+    }
+
+    /**
+     *
+     * @return string
+     * @throws \Exception
+     */
+    public function getPID()
+    {
+        return $this->getCache('pid');
+    }
+
+    /**
+     *
+     * @param mixed $pid
+     * @throws \Exception
+     */
+    public function setPID($pid = '')
+    {
+        $this->setCache('pid', $pid);
+    }
+
+    /**
+     *
+     * @return mixed
+     */
+    public function getLog()
+    {
+        return $this->_log;
+    }
+
+    /**
+     *
+     * @param string $log
+     */
+    public function setLog($log)
+    {
+        $this->_log .= '[' . date('Y-m-d H:i:s') . '][SCENARIO] ' . $log . "\n";
+        if ($this->getConfiguration('logmode', 'default') == 'realtime') {
+            $this->persistLog(true);
+            $this->_log = '';
+        }
+    }
+
+    /**
+     *
+     * @return mixed
+     */
+    public function getRealTrigger()
+    {
+        return $this->_realTrigger;
+    }
+
+    /**
+     *
+     * @param mixed $_realTrigger
+     * @return $this
+     */
+    public function setRealTrigger($_realTrigger)
+    {
+        $this->_realTrigger = $_realTrigger;
+        return $this;
+    }
+
+    /**
+     *
+     * @return string
+     */
+    public function getReturn()
+    {
+        return $this->_return;
+    }
+
+    /**
+     *
+     * @param mixed $_return
+     * @return $this
+     */
+    public function setReturn($_return)
+    {
+        $this->_return = $_return;
+        return $this;
+    }
+
+    /**
+     *
+     * @return array
+     */
+    public function getTags()
+    {
+        return $this->_tags;
+    }
+
+    /**
+     *
+     * @param array $_tags
+     * @return $this
+     */
+    public function setTags($_tags)
+    {
+        $this->_tags = $_tags;
+        return $this;
+    }
+
+    /**
+     *
+     * @return bool
+     */
+    public function getDo()
+    {
+        return $this->_do;
+    }
+
+    /**
+     *
+     * @param bool $_do
+     * @return $this
+     */
+    public function setDo($_do)
+    {
+        $this->_do = $_do;
+        return $this;
+    }
+
+    /**
+     * Nom de la tablea dans la base de données
+     * 
+     * @return string
+     */
+    public function getTableName() 
+    {
+        return 'scenario';
+    }
+
+    /**
+     *
+     * @param string $_key
+     * @param mixed $_default
+     * @return mixed
+     * @throws \Exception
+     */
+    public function getCache($_key = '', $_default = '')
+    {
+        $scenarioCacheAttr = CacheManager::byKey('scenarioCacheAttr' . $this->getId())->getValue();
+        return Utils::getJsonAttr($scenarioCacheAttr, $_key, $_default);
+    }
+
+    /**
+     *
+     * @param mixed $_key
+     * @param mixed $_value
+     * @throws \Exception
+     */
+    public function setCache($_key, $_value = null)
+    {
+        CacheManager::set('scenarioCacheAttr' . $this->getId(), Utils::setJsonAttr(CacheManager::byKey('scenarioCacheAttr' . $this->getId())->getValue(), $_key, $_value));
     }
 
 }
